move exceptions out of main file, move utility functions and PY2/PY3 objects out of main file
